% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{assemble_data}
\alias{assemble_data}
\title{Assemble data for calculating fire metrics}
\usage{
assemble_data(fire_path, from, to, aoi)
}
\arguments{
\item{fire_path}{a character vector of the full file path to the previously
downloaded DBCA Fire History DBCA 060 shape file (see details). File path includes
file extension ".shp"}

\item{from}{numeric representing the starting year for the analysis and
generation of metrics.}

\item{to}{numeric representing the current year for the analysis and
generation of metrics.}

\item{aoi}{the aoi object previously created by using either \code{\link[=user_aoi]{user_aoi()}}
or \code{\link[=DBCA_aoi]{DBCA_aoi()}}}
}
\value{
A list containing the spatially and temporally subsetted Fire History
polygons, the aoi polygon/s, the aoi name and period. Spatial data are of class
SpatVector and are reprojected to Albers GDA2020 (epsg:9473). Note if there is no
data in the Fire History for the aoi, the function returns an error with a
message.
}
\description{
\code{assemble_data} brings together the DBCA Fire History data, the area of
interest (aoi) and the aoi name into a list.
}
\details{
Under the hood, \code{assemble_data} reads from the very large Fire History shape
file only those polygons that intersect the aoi. The user must have previously
downloaded the Fire History shape file \href{https://catalogue.data.wa.gov.au/dataset/dbca-fire-history}{DBCA_Fire_History_DBCA_060}.
The function stores the subset of the Fire History, the aoi and
the aoi name in a named list which will be used by other functions for
calculating various fire metrics. Note all spatial objects have class sf
and data is now reprojected to Albers GDA2020 (epsg:9473).
}
\examples{
\dontrun{
fie_data <- assemble_data(fire_path = "C:/path/to/data/DBCA_Fire_History_DBCA_060.shp",
from = 1988, to = 2022, aoi = aoi)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
