% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{DBCA_aoi}
\alias{DBCA_aoi}
\title{Make a DBCA tenure area of interest}
\usage{
DBCA_aoi(choice, block = FALSE)
}
\arguments{
\item{choice}{a character vector of a single DBCA tenure or forest block name
as determined by previously running \code{\link[=find_tenure]{find_tenure()}} or
\code{\link[=find_block]{find_block()}}.}

\item{block}{logical. Is the choice a forest block? Enables search of correct
shape file of extents.}
}
\value{
A named list containing \code{aoi} (an sf class object) and \code{aoi_name}
(the tenure name of the aoi).
}
\description{
\code{DBCA_aoi} construct a spatial object from DBCA tenure which will be used to
query the DBCA Fire History dataset.
}
\examples{
\dontrun{
aoi <- DBCA_aoi(choice = "Wandoo National Park")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
