% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{find_block}
\alias{find_block}
\title{Find a DBCA forest block name}
\usage{
find_block(like = NULL)
}
\arguments{
\item{like}{default is NULL which returns all DBCA forest block names. Use a
character vector to find matches. Case ambivalent.}
}
\value{
A character vector containing one or many DBCA forest block names.
}
\description{
\code{find_block} helps find the correct DBCA forest block name.
}
\details{
Used with the default parameter, \code{find_block} returns a character vector of
the 400+ full names of all forest blocks in the DBCA estate. A character vector can
be used for a fuzzy search, returning all close matches. Use this function to
obtain the correct search term for \code{\link[=DBCA_aoi]{DBCA_aoi()}}.
}
\examples{
\dontrun{
all_blocks<- find_block(like = NULL)

choice <- find_tenure(like = "jasper")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
