% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{find_tenure}
\alias{find_tenure}
\title{Find a DBCA tenure name}
\usage{
find_tenure(like = NULL)
}
\arguments{
\item{like}{default is NULL which returns all DBCA tenure names. Use a
character vector to find matches. Case ambivalent.}
}
\value{
A character vector containing one or many DBCA tenure names.
}
\description{
\code{find_tenure} helps find the correct tenure name from all of the DBCA estate.
}
\details{
Used with the default parameter, \code{find_tenure} returns a character vector of
the 1000+ full names of all tenure in the DBCA estate. A character vector can
be used for a fuzzy search, returning all close matches. Use this function to
obtain the correct search term for \code{\link[=DBCA_aoi]{DBCA_aoi()}}.
}
\examples{
\dontrun{
all_tenure <- find_tenure(like = NULL)

choice <- find_tenure(like = "wandoo")[1]
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
