% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fire_freq}
\alias{fire_freq}
\title{Metric - Fire Frequency}
\usage{
fire_freq(data, products = TRUE)
}
\arguments{
\item{data}{a data list object created from \code{\link[=assemble_data]{assemble_data()}}}

\item{products}{default TRUE. Renders yslb products to \code{./outputs}}
}
\value{
A list containing the four fire frequency products and if
\code{products = TRUE} a folder containing rendered products.
}
\description{
\code{fire_freq} calculates the fire frequency metrics from the data stored
in the list constructed from the \code{\link[=assemble_data]{assemble_data()}} function.
}
\details{
The function calculates fire frequency both spatially and as area statements
for the area of interest (aoi) and for the time period defined in the previously
constructed data list. The function will create a list object containing four
items:
\itemize{
\item A fire frequency raster object (class SpatRaster)
\item A ggplot object of a 'map' showing the fire frequency raster
\item A data frame of fire frequency area stats
\item A ggplot object of a column plot showing area by fire frequency
}

Using default parameters, the function also outputs the four items to a
created folder ("outputs") in the working directory. All outputs are named by
aoi and time period to differentiate products. Note the fire frequency raster
and product are projected to Albers GDA2020 (epsg:9473).
}
\examples{
\dontrun{
ffreq_list <- fire_freq(data = data_list, products = TRUE)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
