% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fire_interval}
\alias{fire_interval}
\title{Metric - Fire Interval}
\usage{
fire_interval(data, measure = c("min", "max", "mean"), products = TRUE)
}
\arguments{
\item{data}{a data list object created from \code{\link[=assemble_data]{assemble_data()}}}

\item{measure}{a character indicating which fire interval measure to calculate.
Can be one of "min" for minimum, "max" for maximum or "mean" for mean.}

\item{products}{default TRUE. Renders yslb products to \code{./outputs}}
}
\value{
A list containing four fire interval products and if
\code{products = TRUE} a folder containing three rendered products.
}
\description{
\code{fire_interval} calculates  fire interval metrics from the data stored
in the list constructed from the \code{\link[=assemble_data]{assemble_data()}} function.
}
\details{
The function calculates fire intervals both spatially and as area statements
for the area of interest (aoi) and for the time period defined in the previously
constructed data list. The user has a choice of one of 3 interval measures to
choose from. The minimum ("min"), the maximum ("max") or the mean ("mean").

The function will create a list object containing four
items:
\itemize{
\item A fire interval raster object of the chosen measure (class SpatRaster)
\item A ggplot object of a 'map' showing the fire interval raster of the chosen measure
\item A data frame of fire interval area stats of the chosen measure
\item A character vector of the user specified measure
}

Using default parameters, the function also outputs the first three items to a
created folder ("outputs") in the working directory. All outputs are named by
aoi and time period to differentiate products. Note the fire interval raster
and product are projected to Albers GDA2020 (epsg:9473).
}
\examples{
\dontrun{
finterval_list <- fire_interval(data = data_list, measure = "min", products = TRUE)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
