% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{make_wkt}
\alias{make_wkt}
\title{Make well known text string from bounding box of an area of interest}
\usage{
make_wkt(aoi, fh_crs)
}
\arguments{
\item{aoi}{an aoi list object created from running either \code{\link[=user_aoi]{user_aoi()}}
or \code{\link[=DBCA_aoi]{DBCA_aoi()}}.}

\item{fh_crs}{crs of the Fire History input data.}
}
\value{
A WKT representation of the area of interest's bounding box.
}
\description{
\code{make_wkt} translates a bounding box of an area of interest (aoi) into WKT for
use in filtering the DBCA Fire History dataset.
}
\examples{
\dontrun{
wkt_filter <- make_wkt(aoi = aoi)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
