% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{user_aoi}
\alias{user_aoi}
\title{Import a user's area of interest shape file}
\usage{
user_aoi(aoi_path, name)
}
\arguments{
\item{aoi_path}{A character vector of the full file path to the user's shape
file. It will include the name of the shapefile as well as its file extension.}

\item{name}{A character vector name representing the aoi. This name will be
used in titles of outputs and output file names.}
}
\value{
A named list containing \code{aoi} (an sf class object) and \code{aoi_name}
(the name of the aoi).
}
\description{
\code{user_aoi} reads in a shape file from a file path to be used to query the
DBCA Fire History dataset.
}
\details{
Under the hood, \code{user_aoi}, will on reading in the data, repair geometry and
transform the original coordinate reference system to GDA 2020 (Geodetic,
EPSG:7844).
}
\examples{
\dontrun{
aoi <- user_aoi(aoi_filepath = "C:/path/to/data/myshape.shp", name = "Great Shape")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
