% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{yslb}
\alias{yslb}
\title{Metric - Year Since Last Burn}
\usage{
yslb(data, products = TRUE)
}
\arguments{
\item{data}{a data list object created from \code{\link[=assemble_data]{assemble_data()}}}

\item{products}{default TRUE. Renders yslb products to \code{./outputs}}
}
\value{
A list containing the four yslb products and if \code{products = TRUE} a
folder containing rendered products.
}
\description{
\code{yslb} calculates the year since last burn (yslb) metrics from the data stored
in the list constructed from the \code{\link[=assemble_data]{assemble_data()}} function.
}
\details{
The function calculates yslb both spatially and as area statements for the
area of interest (aoi) and for the time period defined in the previously constructed
data list. The function will create a list object containing four items:
\itemize{
\item A yslb raster object (class SpatRaster)
\item A ggplot object of a 'map' showing the yslb raster
\item A data frame of yslb area stats
\item A ggplot object of a column plot showing area by yslb
}

Using default parameters, the function also outputs the four items to a
created folder ("outputs") in the working directory. All outputs are named by
aoi and time period to differentiate products. Note the yslb raster and product
are projected to Albers GDA2020 (epsg:9473).
}
\examples{
\dontrun{
yslb_list <- yslb(data = data_list, products = TRUE)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}
}
